eliminating casting in util

+++ src/com/company/Polynomial.java
@@ -5,13 +5,13 @@ import java.util.List;
 
 public class Polynomial {
 
-    List<Integer> coeffs;
+    List<BigInteger> coeffs;
     BigInteger modulus;
 
-    public void init(List<Integer> coeffs, int mod){
+    public void init(List<BigInteger> coeffs, BigInteger mod){
         //Initializes the polynomial.
         this.coeffs = coeffs;
-        modulus = BigInteger.valueOf(mod);
+        modulus = mod;
 
     }
 
@@ -20,7 +20,7 @@ public class Polynomial {
         BigInteger fx = BigInteger.valueOf(0);
         for (int i = 0; i < coeffs.size(); i++) {
             BigInteger tmp = X.modPow(BigInteger.valueOf(i), modulus);
-            BigInteger tmp2 = tmp.multiply(BigInteger.valueOf(coeffs.get(i))).mod(modulus);
+            BigInteger tmp2 = tmp.multiply(coeffs.get(i)).mod(modulus);
             fx = tmp2.add(fx).mod(modulus);
         }
